requirements.txt
streamlit
PyPDF2
openpyxl
pandas
python-docx
google-generativeai
subprocess
matplotlib
dotenv